+++ CHANGES.md
@@ -3,7 +3,7 @@
 ##### 1.1
 * added BTT - BitConnect currency
 * added [bx.in.th](https://bx.in.th/ref/s9c3HU/) exchange
-* /settings command: language, default currency settings
+* /settings command, only for private chats: language, default currency settings
 * fixes and improvements
 
 ##### 1.0